refactor(autonomi): move structs to top-level
